project1-2 for task1 ( electronics )
Stepper-motor tinkercad code
#include<Stepper.h> const int stepsPerRevolution = 200; //change this to fit the number //of revolutions for your motor

//initialize the stepper library on pins 8 through 11 Stepper mystepper(stepsPerRevolution, 8, 9, 10, 11); int stepCount = 0; // number of steps the motor has taken void setup() { // nothing to do inside the setup } void loop(){ // read the sensor value: int sensorReading = analogRead(A0); // map it to a range from 0 to 100: int motorSpeed = map(sensorReading, 0, 1023, 0, 250); // set the motor speed: if (motorSpeed > 0){ mystepper.setSpeed(motorSpeed); // step 1/100 of a revolution mystepper.step(stepsPerRevolution/100); } }

Stepper_motor_code_Explaination
First, declare the number of revolutions for the stepper motor then, create a function for stepper motor contains for pins (8, 9, 10, 11) and these pins are used as an outouts for control the number of revolutions for the stepper motor. after that, declare the counter that counts the numbee of steps of the stepper motor if the motor speed is greater than 0 which means the motor in the motion phase, the number of steps is executed which are 2 steps = stepsPerRevolution/100 = 200/100=2.